+++ R/repair-names.R
@@ -3,16 +3,16 @@
 #' @description
 #' tibble deals with a few levels of name repair:
 #'   * `minimal` names exist. The `names` attribute is not `NULL`. The name of
-#'   an unnamed element is `""` and never `NA`. Tibbles created by the tibble
-#'   package will have names that are, at least, `minimal`.
+#'     an unnamed element is `""` and never `NA`. Tibbles created by the tibble
+#'     package have names that are, at least, `minimal`.
 #'   * `unique` names are `minimal`, have no duplicates, and are never empty
-#'   (literally, no `""`s).
-#'     - Indexing by name works: `df[["name"]]` extracts exactly one element.
-#'   * `syntactic` names are `unique` and syntactic (see Details for more).
+#'     (literally, no `""`s).
+#'     - All columns can be accessed by name via `df[["name"]]`.
+#'   * `universal` names are `unique` and syntactic (see Details for more).
 #'     - Names work everywhere, without quoting: `df$name` and
 #'     `lm(name1 ~ name2, data = df)` and `dplyr::select(df, name)` all work.
 #'
-#' `syntactic` implies `unique`, `unique` implies `minimal`. These levels are
+#' `universal` implies `unique`, `unique` implies `minimal`. These levels are
 #' nested.
 #'
 #' The `.name_repair` argument of [tibble()] and [as_tibble()] refers to these
@@ -61,17 +61,18 @@
 #' names `unique`, i.e. reconstructing the suffixes. If this interacts poorly
 #' with your names, you should take control of name repair.
 #'
-#' @section `syntactic` names:
+#' @section `universal` names:
 #'
-#' `syntactic` names are `unique` and syntactic, meaning they:
-#'   - Have no duplicates (inherited from `unique`).
-#'   - Consist of letters, numbers, and the dot `.` or underscore `_`
+#' `universal` names are `unique` and syntactic, meaning they:
+#'   * Are never empty (inherited from `unique`).
+#'   * Have no duplicates (inherited from `unique`).
+#'   * Consist of letters, numbers, and the dot `.` or underscore `_`
 #'     characters.
-#'   - Start with a letter or start with the dot `.` not followed by a number.
-#'   - Are not a reserved word, e.g., `if` or `function` or `TRUE`.
-#'   - Are not `...`. Do not have the form `..i`, where `i` is a number.
+#'   * Start with a letter or start with the dot `.` not followed by a number.
+#'   * Are not a [reserved] word, e.g., `if` or `function` or `TRUE`.
+#'   * Are not `...`. Do not have the form `..i`, where `i` is a number.
 #'
-#' If a data frame has `syntactic` names, variable names can be used "as is" in
+#' If a data frame has `universal` names, variable names can be used "as is" in
 #' code. They work well with nonstandard evaluation, e.g., `df$name` works.
 #'
 #' Tibble has a different method of making names syntactic than
@@ -81,10 +82,10 @@
 #' Examples:
 #' ```
 #'  Original names:     ""    "x"    NA     "x"
-#' syntactic names: "...1" "x..2" "...3" "x..4"
+#' universal names: "...1" "x..2" "...3" "x..4"
 #'
 #'   Original names: "(y)"  "_z"  ".2fa"  "FALSE"  "..."  "..3"
-#'  syntactic names: ".y." "._z" "..2fa" ".FALSE" "...." "...6"
+#'  universal names: ".y." "._z" "..2fa" ".FALSE" "...." "...6"
 #' ```
 #'
 #' @param x A vector.
@@ -112,7 +113,7 @@
 #' df$`a 1`
 #'
 #' ## syntactic names are easier to work with, though, and you can request them
-#' df <- tibble(`a 1` = 1, `a 2` = 2, .name_repair = "syntactic")
+#' df <- tibble(`a 1` = 1, `a 2` = 2, .name_repair = "universal")
 #' df$a.1
 #'
 #' ## you can specify your own name repair function
@@ -131,14 +132,14 @@
 NULL
 
 set_repaired_names <- function(x,
-                               .name_repair = c("check_unique", "unique", "syntactic", "minimal")) {
+                               .name_repair = c("check_unique", "unique", "universal", "minimal")) {
   x <- set_minimal_names(x)
   names(x) <- repaired_names(names(x), .name_repair = .name_repair)
   x
 }
 
 repaired_names <- function(name,
-                           .name_repair = c("check_unique", "unique", "syntactic", "minimal")) {
+                           .name_repair = c("check_unique", "unique", "universal", "minimal")) {
 
   if (is_formula(.name_repair, lhs = FALSE)) {
     .name_repair <- as_function(.name_repair)
@@ -156,14 +157,14 @@ repaired_names <- function(name,
       minimal       =     ,
       check_unique  = NULL,
       unique        = unique_names,
-      syntactic     = syntactic_names,
+      universal     = universal_names,
       abort(error_name_repair_arg())
     )
   }
   new_name <- if (is_function(repair_fun)) repair_fun(name) else name
 
   if (is.character(.name_repair) &&
-    .name_repair %in% c("check_unique", "unique", "syntactic")) {
+    .name_repair %in% c("check_unique", "unique", "universal")) {
     check_unique(new_name)
   } else {
     check_minimal(new_name)
@@ -211,27 +212,35 @@ set_unique_names <- function(x, quiet = FALSE) {
   set_names(x, new_names)
 }
 
-syntactic_names <- function(name, quiet = FALSE) {
+universal_names <- function(name, quiet = FALSE) {
   unique_names(name, quiet = quiet, transform = make_syntactic)
 }
 
-set_syntactic_names <- function(x, quiet = FALSE) {
+set_universal_names <- function(x, quiet = FALSE) {
   x <- set_minimal_names(x)
-  new_names <- syntactic_names(names(x), quiet = quiet)
+  new_names <- universal_names(names(x), quiet = quiet)
   set_names(x, new_names)
 }
 
-check_minimal <- function(name) {
+check_names_non_null <- function(name) {
   if (is.null(name)) {
     abort(error_names_must_be_non_null())
   }
+  invisible(name)
+}
 
+check_names_non_na <- function(name) {
   bad_name <- which(is.na(name))
   if (has_length(bad_name)) {
     abort(error_column_must_be_named(bad_name))
   }
 
   invisible(name)
+}
+
+check_minimal <- function(name) {
+  check_names_non_null(name)
+  check_names_non_na(name)
 }
 
 check_minimal_names <- function(x) {
@@ -280,29 +289,34 @@ check_syntactic_names <- function(x) {
 ## makes each individual name syntactic
 ## does not enforce unique-ness
 make_syntactic <- function(name) {
-  name[is.na(name)] <- ""
+  name[is.na(name)]       <- ""
+  name[name == ""]        <- "."
+  name[name == "..."]     <- "...."
+  name <- sub("^_", "._", name)
+
   new_name <- make.names(name)
 
   X_prefix <- grepl("^X", new_name) & !grepl("^X", name)
-  new_name[X_prefix] <- gsub("^X", "", new_name[X_prefix])
+  new_name[X_prefix] <- sub("^X", "", new_name[X_prefix])
 
-  dot_suffix <- nchar(new_name) == (nchar(name) + 1) &
-    grepl("[.]$", new_name) & !grepl("[.]$", name)
-  new_name[dot_suffix] <- gsub("[.]$", "", new_name[dot_suffix])
-  new_name[dot_suffix] <- paste0(".", new_name[dot_suffix])
+  dot_suffix <- which(new_name == paste0(name, "."))
+  new_name[dot_suffix] <- sub("^(.*)[.]$", ".\\1", new_name[dot_suffix])
   ## illegal characters have been replaced with '.' via make.names()
   ## however, we have:
   ##   * declined its addition of 'X' prefixes
   ##   * turned its '.' suffixes to '.' prefixes
 
-  ## ".i" --> "..i", so it's caught in next step
-  new_name <- gsub("^([.][1-9][0-9]*)$", ".\\1", new_name)
-
-  new_name <- gsub(
-    "^(|[_].*|[.][.][.]|[.][.][1-9][0-9]*|[.][0-9].*)$",
-    ".\\1",
-    new_name
+  regex <- paste0(
+    "^(?<leading_dots>[.]{0,2})",
+    "(?<numbers>[0-9]*)",
+    "(?<leftovers>[^0-9]?.*$)"
   )
+
+  re <- re_match(new_name, pattern = regex)
+  needs_dots <- which(re$numbers != "")
+  needs_third_dot <- (re$leftovers[needs_dots] == "")
+  re$leading_dots[needs_dots] <- ifelse(needs_third_dot, "...", "..")
+  new_name <- paste0(re$leading_dots, re$numbers, re$leftovers)
 
   new_name
 }
@@ -342,10 +356,10 @@ describe_repair <- function(orig_name, name) {
 #' `as_tibble()` did not do this.
 #'
 #' From tibble v1.5.0, the `.name_repair` argument gives direct access to three
-#' specific levels of name repair: `minimal`, `unique`, and `syntactic`. We
+#' specific levels of name repair: `minimal`, `unique`, and `universal`. We
 #' recommend that new code use this instead of `tidy_names()`,
 #' `set_tidy_names()`, or `repair_names()`. After a period of use, the repair
-#' stategies behind `minimal`, `unique`, and `syntactic` are likely to be
+#' stategies behind `minimal`, `unique`, and `universal` are likely to be
 #' exposed in standalone functions and this could affect the behaviour of
 #' `tidy_names()`. `repair_names()` should be considered deprecated.
 #'
@@ -355,7 +369,7 @@ describe_repair <- function(orig_name, name) {
 #' df <- tibble(...)
 #' set_tidy_names(df, syntactic = FALSE)
 #'
-#' tibble(..., `.name_repair = "syntactic"`)
+#' tibble(..., `.name_repair = "universal"`)
 #' ## is preferred to
 #' df <- tibble(...)
 #' set_tidy_names(df, syntactic = TRUE)
@@ -363,12 +377,12 @@ describe_repair <- function(orig_name, name) {
 #'
 #' @param syntactic Should names be made syntactically valid? If `FALSE`, uses
 #'   same logic as `.name_repair = "unique"`. If `TRUE`, uses same logic as
-#'   `.name_repair = "syntactic"`.
+#'   `.name_repair = "universal"`.
 #' @export
 #' @rdname name-repair
 tidy_names <- function(name, syntactic = FALSE, quiet = FALSE) {
   if (syntactic) {
-    syntactic_names(name, quiet = quiet)
+    universal_names(name, quiet = quiet)
   } else {
     unique_names(name, quiet = quiet)
   }