+++ R/utils.r
@@ -17,11 +17,19 @@ has_nonnull_names <- function(x) {
 
 set_class <- `class<-`
 
+check_no_dim <- function(x) {
+  if (is_atomic(x) && has_dim(x)) {
+    abort(error_1d_array_column())
+  }
+  invisible(x)
+}
+
 strip_dim <- function(x) {
   if (is.matrix(x)) {
     rownames(x) <- NULL
   } else if (is.data.frame(x)) {
     x <- remove_rownames(x)
+    x[] <- map(x, strip_dim)
   } else if (is_atomic(x) && has_dim(x)) {
     # Careful update only if necessary, to avoid copying which is checked by
     # the "copying" test in dplyr
@@ -72,4 +80,42 @@ tick_if_needed <- function(x) {
   needs_ticks <- !is_syntactic(x)
   x[needs_ticks] <- tick(x[needs_ticks])
   x
+}
+
+## from rematch2, except we don't add tbl_df or tbl classes to the return value
+re_match <- function(text, pattern, perl = TRUE, ...) {
+
+  stopifnot(is.character(pattern), length(pattern) == 1, !is.na(pattern))
+  text <- as.character(text)
+
+  match <- regexpr(pattern, text, perl = perl, ...)
+
+  start  <- as.vector(match)
+  length <- attr(match, "match.length")
+  end    <- start + length - 1L
+
+  matchstr <- substring(text, start, end)
+  matchstr[ start == -1 ] <- NA_character_
+
+  res <- data.frame(
+    stringsAsFactors = FALSE,
+    .text = text,
+    .match = matchstr
+  )
+
+  if (!is.null(attr(match, "capture.start"))) {
+
+    gstart  <- attr(match, "capture.start")
+    glength <- attr(match, "capture.length")
+    gend    <- gstart + glength - 1L
+
+    groupstr <- substring(text, gstart, gend)
+    groupstr[ gstart == -1 ] <- NA_character_
+    dim(groupstr) <- dim(gstart)
+
+    res <- cbind(groupstr, res, stringsAsFactors = FALSE)
+  }
+
+  names(res) <- c(attr(match, "capture.names"), ".text", ".match")
+  res
 }