+++ tests/testthat/test-data-frame.R
@@ -202,21 +202,22 @@ test_that("Can convert named atomic vectors to data frame", {
 
 test_that("as_tibble() checks for `unique` names by default (#278)", {
   l1 <- list(1:10)
-  l2 <- list(x = 1, 2)
-
-  df <- tibble(a = 1, b = 2)
-  names(df) <- c("", NA)
-
   expect_error(
     as_tibble(l1),
     error_column_must_be_named(1, repair = TRUE),
     fixed = TRUE
   )
+
+  l2 <- list(x = 1, 2)
   expect_error(
     as_tibble(l2),
     error_column_must_be_named(2, repair = TRUE),
     fixed = TRUE
   )
+
+  df <- list(a = 1, b = 2)
+  names(df) <- c("", NA)
+  df <- new_tibble(df, nrow = 1)
   expect_error(
     as_tibble(df),
     error_column_must_be_named(1:2, repair = TRUE),
@@ -239,11 +240,11 @@ test_that("as_tibble() implements unique names", {
   expect_equal(names(invalid_df), unique_names(rep("", 3)))
 })
 
-test_that("as_tibble() implements syntactic names", {
-  invalid_df <- as_tibble(list(3, 4, 5), .name_repair = "syntactic")
+test_that("as_tibble() implements universal names", {
+  invalid_df <- as_tibble(list(3, 4, 5), .name_repair = "universal")
   expect_equal(length(invalid_df), 3)
   expect_equal(nrow(invalid_df), 1)
-  expect_equal(names(invalid_df), syntactic_names(rep("", 3)))
+  expect_equal(names(invalid_df), universal_names(rep("", 3)))
 })
 
 
@@ -263,6 +264,129 @@ test_that("as_tibble() implements custom name repair", {
   expect_identical(invalid_df_purrr, invalid_df)
 })
 
+test_that("as_tibble.matrix() supports .name_repair", {
+  x <- matrix(1:6, nrow = 3)
+
+  expect_error(
+    as_tibble(x),
+    error_column_must_be_named(1:2)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    rep("", 2)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    paste0("...", 1:2)
+  )
+
+  x <- matrix(1:6, nrow = 3, dimnames = list(x = LETTERS[1:3], y = c("if", "when")))
+
+  expect_identical(
+    names(as_tibble(x)),
+    c("if", "when")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    c("if", "when")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    c(".if", "when")
+  )
+})
+
+test_that("as_tibble.poly() supports .name_repair", {
+  x <- poly(1:6, 3)
+
+  expect_identical(
+    names(as_tibble(x)),
+    as.character(1:3)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    as.character(1:3)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    paste0("...", 1:3)
+  )
+})
+
+test_that("as_tibble.table() supports .name_repair", {
+  x <- table(a = c(1, 1, 1, 2, 2, 2), a = c(3, 4, 5, 3, 4, 5))
+
+  expect_error(
+    as_tibble(x),
+    error_column_names_must_be_unique("a")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    c("a", "a", "n")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    c("a..1", "a..2", "n")
+  )
+
+  x <- table("if" = c(1, 1, 1, 2, 2, 2), "when" = c(3, 4, 5, 3, 4, 5))
+
+  expect_identical(
+    names(as_tibble(x)),
+    c("if", "when", "n")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    c("if", "when", "n")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    c(".if", "when", "n")
+  )
+
+  x <- table("m" = c(1, 1, 1, 2, 2, 2), "n" = c(3, 4, 5, 3, 4, 5))
+
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    c("m", "n", "n")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    c("m", "n..2", "n..3")
+  )
+})
+
+test_that("as_tibble.ts() supports .name_repair", {
+  x <- ts(matrix(rnorm(6), nrow = 3), start = c(1961, 1), frequency = 12, names = NULL)
+
+  expect_error(
+    names(as_tibble(x)),
+    error_column_must_be_named(1:2)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    rep("", 2)
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    paste0("...", 1:2)
+  )
+
+  x <- ts(matrix(rnorm(6), nrow = 3), start = c(1961, 1), frequency = 12, names = c("if", "when"))
+
+  expect_identical(
+    names(as_tibble(x)),
+    c("if", "when")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "minimal")),
+    c("if", "when")
+  )
+  expect_identical(
+    names(as_tibble(x, .name_repair = "universal")),
+    c(".if", "when")
+  )
+})
 
 test_that("as_tibble() can convert row names", {
   df <- data.frame(a = 1:3, b = 2:4, row.names = letters[5:7])
@@ -296,17 +420,6 @@ test_that("as_data_frame is an alias of as_tibble", {
 test_that("as.tibble is an alias of as_tibble", {
   expect_identical(as.tibble(NULL), as_tibble(NULL))
 })
-
-# new_tibble -------------------------------------------------------------
-
-test_that("new_tibble can specify nrow,", {
-  expect_error(
-    new_tibble(list(x = 1:2, y = 1:3), nrow = 4),
-    error_inconsistent_cols(4, c("x", "y"), c(2, 3), "`nrow` argument"),
-    fixed = TRUE
-  )
-})
-
 
 # Validation --------------------------------------------------------------