+++ tests/testthat/test-matrix.R
@@ -2,7 +2,7 @@ context("matrix")
 
 test_that("correct rows and cols", {
   x <- matrix(1:6, nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
 
   expect_equal(dim(out), c(2, 3))
 })
@@ -15,11 +15,14 @@ test_that("preserves col names", {
   expect_equal(names(out), c("a", "b"))
 })
 
-test_that("creates col names", {
+test_that("creates col names with name repair", {
   x <- matrix(1:4, nrow = 2)
 
-  out <- as_tibble(x)
-  expect_equal(names(out), c("V1", "V2"))
+  out <- as_tibble(x, .name_repair = "unique")
+  expect_equal(names(out), c("..1", "..2"))
+
+  out <- as_tibble(x, .name_repair = "universal")
+  expect_equal(names(out), c("...1", "...2"))
 })
 
 test_that("preserves attributes except dim and names", {
@@ -44,31 +47,24 @@ test_that("properly handles poly class (#110)", {
 
 test_that("handles atomic vectors", {
   x <- matrix(TRUE, nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
   expect_equal(out[[1]], c(TRUE, TRUE))
 
   x <- matrix(1L, nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
   expect_equal(out[[1]], c(1L, 1L))
 
   x <- matrix(1.5, nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
   expect_equal(out[[1]], c(1.5, 1.5))
 
   x <- matrix("a", nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
   expect_equal(out[[1]], c("a", "a"))
 
   x <- matrix(complex(real = 1, imag = 2), nrow = 2)
-  out <- as_tibble(x)
+  out <- as_tibble(x, .name_repair = "minimal")
   expect_equal(out[[1]], as.vector(x))
-})
-
-test_that("auto-assigning names", {
-  expect_identical(
-    as_tibble(diag(3L)),
-    as_tibble(as.data.frame(diag(3L)))
-  )
 })
 
 test_that("forwarding to as.data.frame() for ts objects (#184)", {
@@ -110,7 +106,7 @@ test_that("converting from matrix supports storing row names in a column", {
 test_that("converting from matrix throws an error if user turns missing row names into column", {
   x <- matrix(1:30, 6, 5)
   expect_error(
-    as_tibble(x, rownames = "id"),
+    as_tibble(x, rownames = "id", .name_repair = "minimal"),
     error_as_tibble_needs_rownames(),
     fixed = TRUE
   )