+++ tests/testthat/test-tbl-df.R
@@ -245,6 +245,33 @@ test_that("[.tbl_df ignores drop argument (with warning) without j argument (#30
 })
 
 
+test_that("[.tbl_df is careful about attributes (#155)", {
+  df <- tibble(x = 1:2, y = x)
+  attr(df, "along for the ride") <- "still here"
+
+  expect_identical(attr(df[names(df)], "along for the ride"), "still here")
+  expect_identical(attr(df["x"], "along for the ride"), "still here")
+  expect_identical(attr(df[1:2], "along for the ride"), "still here")
+  expect_identical(attr(df[2], "along for the ride"), "still here")
+  expect_identical(attr(df[c(TRUE, FALSE)], "along for the ride"), "still here")
+  expect_identical(attr(df[, names(df)], "along for the ride"), "still here")
+  expect_identical(attr(df[, "x"], "along for the ride"), "still here")
+  expect_identical(attr(df[, 1:2], "along for the ride"), "still here")
+  expect_identical(attr(df[, 2], "along for the ride"), "still here")
+  expect_identical(attr(df[, c(TRUE, FALSE)], "along for the ride"), "still here")
+  expect_identical(attr(df[1, names(df)], "along for the ride"), "still here")
+  expect_identical(attr(df[1, "x"], "along for the ride"), "still here")
+  expect_identical(attr(df[1, 1:2], "along for the ride"), "still here")
+  expect_identical(attr(df[1, 2], "along for the ride"), "still here")
+  expect_identical(attr(df[1, c(TRUE, FALSE)], "along for the ride"), "still here")
+
+  expect_identical(attr(df[1:2, ], "along for the ride"), "still here")
+  expect_identical(attr(df[-1, ], "along for the ride"), "still here")
+
+  expect_identical(attr(df[, ], "along for the ride"), "still here")
+  expect_identical(attr(df[], "along for the ride"), "still here")
+})
+
 # [[ ----------------------------------------------------------------------
 
 test_that("[[.tbl_df ignores exact argument", {
@@ -315,9 +342,10 @@ test_that("is_tibble", {
 
 # new_tibble --------------------------------------------------------------
 
-test_that("new_tibble", {
+test_that("new_tibble() with deprecated subclass argument", {
   tbl <- new_tibble(
     data.frame(a = 1:3),
+    names = "b",
     attr1 = "value1",
     attr2 = 2,
     nrow = 3,
@@ -329,10 +357,34 @@ test_that("new_tibble", {
   expect_equal(
     unclass(tbl),
     structure(
-      list(a = 1:3),
+      list(b = 1:3),
       attr1 = "value1",
       attr2 = 2,
-      .Names = "a",
+      .Names = "b",
+      row.names = .set_row_names(3L)
+    )
+  )
+})
+
+test_that("new_tibble() with new class argument", {
+  tbl <- new_tibble(
+    data.frame(a = 1:3),
+    names = "b",
+    attr1 = "value1",
+    attr2 = 2,
+    nrow = 3,
+    class = "nt"
+  )
+
+  # Can't compare directly due to dplyr:::all.equal.tbl_df()
+  expect_identical(class(tbl), c("nt", "tbl_df", "tbl", "data.frame"))
+  expect_equal(
+    unclass(tbl),
+    structure(
+      list(b = 1:3),
+      attr1 = "value1",
+      attr2 = 2,
+      .Names = "b",
       row.names = .set_row_names(3L)
     )
   )
@@ -343,14 +395,28 @@ test_that("new_tibble checks", {
   expect_identical(new_tibble(list(), nrow = 5), tibble(.rows = 5))
   expect_identical(new_tibble(list(a = 1:3, b = 4:6), nrow = 3), tibble(a = 1:3, b = 4:6))
   expect_error(
+    new_tibble(1:3, nrow = 1),
+    error_new_tibble_must_be_list(),
+    fixed = TRUE
+  )
+  expect_error(
+    new_tibble(list(1)),
+    error_new_tibble_needs_nrow(),
+    fixed = TRUE
+  )
+  expect_error(
+    new_tibble(list(1), nrow = NULL),
+    error_new_tibble_needs_nrow(),
+    fixed = TRUE
+  )
+  expect_error(
     new_tibble(list(1), nrow = 1),
     error_names_must_be_non_null(repair = FALSE),
     fixed = TRUE
   )
   expect_error(
     new_tibble(set_names(list(1), NA_character_), nrow = 1),
-    error_column_must_be_named(1, repair = FALSE),
-    fixed = TRUE
+    NA
   )
   expect_error(
     new_tibble(set_names(list(1), ""), nrow = 1),
@@ -358,8 +424,7 @@ test_that("new_tibble checks", {
   )
   expect_error(
     new_tibble(list(a = 1, b = 2:3), nrow = 1),
-    error_inconsistent_cols(1, c("a", "b"), 1:2, "`nrow` argument"),
-    fixed = TRUE
+    NA
   )
   expect_error(
     new_tibble(
@@ -367,5 +432,20 @@ test_that("new_tibble checks", {
       nrow = 1
     ),
     NA
+  )
+})
+
+
+test_that("validate_tibble() checks", {
+  expect_error(
+    validate_tibble(new_tibble(list(a = 1, b = 2:3), nrow = 1)),
+    error_inconsistent_cols(1, c("a", "b"), 1:2, "`nrow` argument"),
+    fixed = TRUE
+  )
+
+  expect_error(
+    validate_tibble(new_tibble(list(a = array(1:3)), nrow = 3)),
+    error_1d_array_column(),
+    fixed = TRUE
   )
 })